Move RAG schemas to schema module
