+++ routers/rag.py
@@ -1,63 +1,18 @@
 from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel, Field
 
 from services.rag_service import answer_question, search_documents
+from schemas.rag import (
+    RagSearchRequest, 
+    RagSearchResponse,
+    RagSearchResultResponse,
+    RagAskRequest,
+    RagAskResponse,
+    RagSourceResponse,
+)
 
-from typing import Optional
 
 router = APIRouter(prefix="/rag", tags=["RAG"]) # 创建一个 APIRouter 实例，设置前缀为 "/rag"，所有在这个 router 中定义的路由都会自动加上这个前缀。同时给这个 router 打上 "RAG" 标签，方便在 API 文档中分类显示。
 
-class RagSearchRequest(BaseModel):
-    '''
-    用户发给 API 的请求
-    '''
-    query: str = Field(..., min_length=1)   
-    top_k: int = Field(default=3, ge=1, le=10)  # 大于等于1，小于等于10
-
-class RagSearchResultResponse(BaseModel):
-    '''
-    一条检索结果
-    '''
-    rank: int
-    document_id: str
-    chunk_id: str
-    title: str
-    source_path: str
-    chunk_index: int
-    distance: float
-    preview: str
-
-class RagSearchResponse(BaseModel):
-    '''
-    整个接口返回结果
-    '''
-    query: str
-    top_k: int
-    total_hits: int
-    results: list[RagSearchResultResponse]
-
-class RagAskRequest(BaseModel): # 用户问问题时传进来的请求体
-    question: str = Field(..., min_length=1)
-    top_k: int = Field(default=3, ge=1, le=10)
-    max_distance: float = Field(default=0.9, gt=0)
-
-class RagSourceResponse(BaseModel): # sources 里的单条引用来源
-    rank: int
-    document_id: str
-    chunk_id: str
-    title: str
-    source_path: str
-    chunk_index: int
-    distance: float
-    preview: str
-
-
-class RagAskResponse(BaseModel):    # 整个问答接口返回体
-    question: str
-    answer: str
-    retrieval_status: str
-    top_distance: Optional[float] = None
-    sources: list[RagSourceResponse]
 
 def make_preview(text: str, max_length: int = 200) -> str:
     '''
@@ -73,6 +28,9 @@ def make_preview(text: str, max_length: int = 200) -> str:
 
 @router.post("/search", response_model=RagSearchResponse)
 def rag_search(request: RagSearchRequest):
+    '''
+    根据用户的查询语句，搜索相关的文档片段
+    '''
     try:
         results = search_documents(
             query=request.query,
